requirements.txt
packaging==23.2
Pillow==10.0.1
pytesseract==0.3.10
